pyproject.toml
[tool.poetry]
name = "toxicity-kaggle"
version = "0.1.0"
description = "Classify toxic online comments challenge from Kaggle"
authors = ["Dudu Lasry <dudu.lasry@viz.ai>"]
readme = "README.md"
packages = [{include = "toxicity_kaggle"}]

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.25.1"
torch = "1.12.1"
pandas = "^1.5.2"
numpy = "^1.24.1"
matplotlib = "^3.6.2"
ipykernel = "^6.19.4"
kaggle = "^1.5.12"
datasets = "^2.8.0"
wordcloud = "^1.8.2.2"
nltk = "^3.8"
hydra-core = "^1.3.1"
clearml = "^1.9.0"
scikit-learn = "^1.2.0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
